test(e2e): fold http: config spec into e2e-critical-a's shared HA instance
